+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- PHP Error handling configuration
+- *MySQL* server and database
+
+### Changed
+- Provisioning script to function based one (`src`) to improve flexibility
+- Default `vhost` moved to `conf` folder
+
 ## [0.1.0] - 2016-09-11
 ### Added
 - Base setup